move weightedsplitting, robustsplitting, a2a and p2p to loss.mri submodule
